error correction in weeks assessment and addition of plasma analysis
as a first point, the assessment of the weeks in the function ncases has been corrected. This has an influence on the crossvalidation scores. As a second point, a for cycle has been implemented in the analysis file, allowing for the analysis of the plasma files that have recently been received. A corresponding option for the choice of plasma (p), all (plasma + liver -> a) and liver (default) has been added